feat(s3): turn driver + end-of-game + client plumbing
Server:
- New TurnDriver (setTimeout-backed, per-room, 30s default from
  room.settings.phase_timer_sec)
- Room.ts: remove per-action timer restart; add terminated flag,
  shutdown(), isTerminated(), getTurnDeadline(); expand attach() to
  {send, close} with backward-compat overload; broadcast turn_advance
  on currentPlayerIdx change; microtask-queue runFallback on AI
  turn-start; fire onGameOver once per state.winner transition
- registry.ts: own TurnDriver; onTurnAdvance / onTurnExpire handlers;
  setOnGameOver setter (avoids circular import); per-phase
  forcedActionFor for human force-advance on timer expiry
- ws/index.ts: pass {send, close} to room.attach; populate
  welcome.turnDeadlineMs
- index.ts: wire handleGameOver via registry.setOnGameOver

New endGame.ts: handleGameOver broadcasts game_over → calls end_game
RPC → 500ms flush → room.shutdown → registry.delete.

Client:
- useRoomDispatcher: handle turn_advance + game_over; propagate
  turnDeadlineMs from welcome
- PlayRoom: useTurnClock (250ms rerender); live clock; 3s auto-redirect
  on game_over; reconnect fallback when state='finished'

Shared protocol: ServerTurnAdvanceSchema, ServerGameOverSchema,
optional turnDeadlineMs on welcome (additive with S3 Agent 1).

24 new server tests + 2 new web tests pass.

+++ packages/shared/src/protocol.ts
@@ -85,11 +85,13 @@ export const ServerWelcomeSchema = z.object({
   hash: HashSchema,
   seq: SeqSchema,
   /**
-   * Absolute epoch ms at which the current seat's turn budget expires.
+   * Absolute epoch-ms at which the current seat's turn budget expires.
    * Optional because rooms that haven't launched yet (setup-claim without
-   * TurnDriver) won't emit it. Clients that see it start a countdown.
+   * TurnDriver) won't emit it, and reconnects that arrive between turns or
+   * after the game ended won't either. Clients that see it start a
+   * countdown.
    */
-  turnDeadlineMs: z.number().optional(),
+  turnDeadlineMs: z.number().int().optional(),
 });
 
 export const ServerAppliedSchema = z.object({
@@ -132,20 +134,22 @@ export const ServerErrorSchema = z.object({
 });
 
 /**
- * Fired whenever the server changes the active seat. Clients use it to
- * restart their per-turn countdown and optionally re-surface whose turn it
- * is. `turnNumber` matches `GameState.turn` post-advance.
+ * Fired whenever the server changes the active seat (`state.currentPlayerIdx`).
+ * Clients use it to restart their per-turn countdown and optionally
+ * re-surface whose turn it is. `turnNumber` matches `GameState.turn`
+ * post-advance.
  */
 export const ServerTurnAdvanceSchema = z.object({
   type: z.literal('turn_advance'),
   currentSeatIdx: SeatIdxSchema,
   turnNumber: z.number().int().min(0),
-  deadlineMs: z.number(),
+  deadlineMs: z.number().int(),
 });
 
 /**
- * Terminal frame when the engine reports a winner. `winnerUserId` is null
- * for AI winners; `winnerSeatIdx` is null only if the winner cannot be
+ * Terminal frame emitted exactly once when the engine sets `state.winner`.
+ * The server closes sockets ~500ms after broadcasting. `winnerUserId` is
+ * null for AI winners; `winnerSeatIdx` is null only if the winner cannot be
  * resolved back to a seat (defensive — shouldn't happen in practice).
  */
 export const ServerGameOverSchema = z.object({
